[autopilot] support generated autopilot, based on rotorcraft firmware
By default the original static autopilot is used. A config flag can
enable the use of the generated AP code.
A basic autopilot description is provided (4 modes + failsafe modes).
The server is capable of using the list of generated mode to properly
display mode names.
Tested in NAV and GUIDED mode in sim, and direct attitude control on
real aircraft.